update from origin

+++ utilcode/src/main/java/com/blankj/utilcode/util/CrashUtils.java
@@ -1,5 +1,6 @@
 package com.blankj.utilcode.util;
 
+import android.annotation.SuppressLint;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
@@ -14,7 +15,8 @@ import java.lang.Thread.UncaughtExceptionHandler;
 import java.text.Format;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.Locale;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 /**
  * <pre>
@@ -26,23 +28,29 @@ import java.util.Locale;
  */
 public final class CrashUtils {
 
-    private static boolean mInitialized;
-    private static String  defaultDir;
-    private static String  dir;
-    private static String  versionName;
-    private static int     versionCode;
+    private static String defaultDir;
+    private static String dir;
+    private static String versionName;
+    private static int    versionCode;
+
+    private static ExecutorService sExecutor;
 
     private static final String FILE_SEP = System.getProperty("file.separator");
-    private static final Format FORMAT   = new SimpleDateFormat("MM-dd HH-mm-ss", Locale.getDefault());
+    @SuppressLint("SimpleDateFormat")
+    private static final Format FORMAT   = new SimpleDateFormat("MM-dd HH-mm-ss");
 
     private static final String CRASH_HEAD;
 
     private static final UncaughtExceptionHandler DEFAULT_UNCAUGHT_EXCEPTION_HANDLER;
     private static final UncaughtExceptionHandler UNCAUGHT_EXCEPTION_HANDLER;
 
+    private static OnCrashListener sOnCrashListener;
+
     static {
         try {
-            PackageInfo pi = Utils.getContext().getPackageManager().getPackageInfo(Utils.getContext().getPackageName(), 0);
+            PackageInfo pi = Utils.getApp()
+                    .getPackageManager()
+                    .getPackageInfo(Utils.getApp().getPackageName(), 0);
             if (pi != null) {
                 versionName = pi.versionName;
                 versionCode = pi.versionCode;
@@ -51,11 +59,11 @@ public final class CrashUtils {
             e.printStackTrace();
         }
 
-        CRASH_HEAD = "\n************* Crash Log Head ****************" +
+        CRASH_HEAD = "************* Crash Log Head ****************" +
                 "\nDevice Manufacturer: " + Build.MANUFACTURER +// 设备厂商
                 "\nDevice Model       : " + Build.MODEL +// 设备型号
                 "\nAndroid Version    : " + Build.VERSION.RELEASE +// 系统版本
-                "\nAndroid SDK        : " + Build.VERSION.SDK_INT +// SDK版本
+                "\nAndroid SDK        : " + Build.VERSION.SDK_INT +// SDK 版本
                 "\nApp VersionName    : " + versionName +
                 "\nApp VersionCode    : " + versionCode +
                 "\n************* Crash Log Head ****************\n\n";
@@ -66,15 +74,25 @@ public final class CrashUtils {
             @Override
             public void uncaughtException(final Thread t, final Throwable e) {
                 if (e == null) {
-                    android.os.Process.killProcess(android.os.Process.myPid());
-                    System.exit(0);
+                    if (DEFAULT_UNCAUGHT_EXCEPTION_HANDLER != null) {
+                        DEFAULT_UNCAUGHT_EXCEPTION_HANDLER.uncaughtException(t, null);
+                    } else {
+                        android.os.Process.killProcess(android.os.Process.myPid());
+                        System.exit(1);
+                    }
                     return;
+                }
+                if (sOnCrashListener != null) {
+                    sOnCrashListener.onCrash(e);
                 }
                 Date now = new Date(System.currentTimeMillis());
                 String fileName = FORMAT.format(now) + ".txt";
                 final String fullPath = (dir == null ? defaultDir : dir) + fileName;
                 if (!createOrExistsFile(fullPath)) return;
-                new Thread(new Runnable() {
+                if (sExecutor == null) {
+                    sExecutor = Executors.newSingleThreadExecutor();
+                }
+                sExecutor.execute(new Runnable() {
                     @Override
                     public void run() {
                         PrintWriter pw = null;
@@ -95,7 +113,7 @@ public final class CrashUtils {
                             }
                         }
                     }
-                }).start();
+                });
                 if (DEFAULT_UNCAUGHT_EXCEPTION_HANDLER != null) {
                     DEFAULT_UNCAUGHT_EXCEPTION_HANDLER.uncaughtException(t, e);
                 }
@@ -109,47 +127,74 @@ public final class CrashUtils {
 
     /**
      * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
-     *
-     * @return {@code true}: 初始化成功<br>{@code false}: 初始化失败
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      */
-    public static boolean init() {
-        return init("");
+    public static void init() {
+        init("");
     }
 
     /**
      * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      *
      * @param crashDir 崩溃文件存储目录
-     * @return {@code true}: 初始化成功<br>{@code false}: 初始化失败
      */
-    public static boolean init(@NonNull final File crashDir) {
-        return init(crashDir.getAbsolutePath() + FILE_SEP);
+    public static void init(@NonNull final File crashDir) {
+        init(crashDir.getAbsolutePath(), null);
     }
 
     /**
      * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      *
      * @param crashDir 崩溃文件存储目录
-     * @return {@code true}: 初始化成功<br>{@code false}: 初始化失败
      */
-    public static boolean init(final String crashDir) {
+    public static void init(final String crashDir) {
+        init(crashDir, null);
+    }
+
+    /**
+     * 初始化
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param onCrashListener 崩溃监听事件
+     */
+    public static void init(final OnCrashListener onCrashListener) {
+        init("", onCrashListener);
+    }
+
+    /**
+     * 初始化
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param crashDir        崩溃文件存储目录
+     * @param onCrashListener 崩溃监听事件
+     */
+    public static void init(@NonNull final File crashDir, final OnCrashListener onCrashListener) {
+        init(crashDir.getAbsolutePath(), onCrashListener);
+    }
+
+    /**
+     * 初始化
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param crashDir        崩溃文件存储目录
+     * @param onCrashListener 崩溃监听事件
+     */
+    public static void init(final String crashDir, final OnCrashListener onCrashListener) {
         if (isSpace(crashDir)) {
             dir = null;
         } else {
-            dir = crashDir.endsWith(FILE_SEP) ? dir : dir + FILE_SEP;
+            dir = crashDir.endsWith(FILE_SEP) ? crashDir : crashDir + FILE_SEP;
         }
-        if (mInitialized) return true;
         if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState())
-                && Utils.getContext().getExternalCacheDir() != null)
-            defaultDir = Utils.getContext().getExternalCacheDir() + FILE_SEP + "crash" + FILE_SEP;
+                && Utils.getApp().getExternalCacheDir() != null)
+            defaultDir = Utils.getApp().getExternalCacheDir() + FILE_SEP + "crash" + FILE_SEP;
         else {
-            defaultDir = Utils.getContext().getCacheDir() + FILE_SEP + "crash" + FILE_SEP;
+            defaultDir = Utils.getApp().getCacheDir() + FILE_SEP + "crash" + FILE_SEP;
         }
+        sOnCrashListener = onCrashListener;
         Thread.setDefaultUncaughtExceptionHandler(UNCAUGHT_EXCEPTION_HANDLER);
-        return mInitialized = true;
     }
 
     private static boolean createOrExistsFile(final String filePath) {
@@ -176,5 +221,9 @@ public final class CrashUtils {
             }
         }
         return true;
+    }
+
+    public interface OnCrashListener {
+        void onCrash(Throwable e);
     }
 }
